requirements.txt
# Streamlit
streamlit>=1.28.0

# LangChain (OpenRouter-compatible via ChatOpenAI)
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# Vector store
chromadb>=0.5.0
langchain-chroma>=0.1.0

# PDF ingestion
pymupdf>=1.24.0          # import fitz

# Knowledge graph
networkx>=3.0
spacy>=3.7.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# Web search fallback (CRAG)
tavily-python>=0.3.0

# Reranking (optional cross-encoder)
sentence-transformers>=2.7.0

# BM25 (used in Fusion RAG)
rank-bm25>=0.2.2

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.7.0
tqdm>=4.0.0
pandas>=2.0.0
numpy>=1.26.0
